requirements.txt
flask
flask-cors
torch
torchvision
numpy
opencv-python
pillow
matplotlib
scikit-learn
